aio fixes and more int -> ssize_t
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@17409 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ sliod/slvr.h
@@ -193,9 +193,9 @@ struct slvr_ref *
 void	slvr_cache_init(void);
 void	slvr_clear_inuse(struct slvr_ref *, int, uint32_t);
 int	slvr_do_crc(struct slvr_ref *);
-int	slvr_fsbytes_wio(struct sli_iocb_set **, struct slvr_ref *,
+ssize_t	slvr_fsbytes_wio(struct sli_iocb_set **, struct slvr_ref *,
 	    uint32_t, uint32_t);
-int	slvr_io_prep(struct pscrpc_request *, struct sli_iocb_set **,
+ssize_t	slvr_io_prep(struct pscrpc_request *, struct sli_iocb_set **,
 	    struct slvr_ref *, uint32_t, uint32_t, enum rw);
 void	slvr_repl_prep(struct slvr_ref *, int);
 void	slvr_rio_done(struct slvr_ref *);
